backend/requirements.txt
fastapi
uvicorn
python-multipart

bcrypt
PyJWT

requests

motor
pymongo

python-dotenv

google-generativeai